Add support for CSV output format.
When `--csv` argument is used, the utility produces CSV format.
Nested top level elements (arrays, objects) are formatted as JSON.

Using the option alone saves some time already:

```
PS C:\Users\mispecto\Projects\azure-kusto-parquet-conv> Measure-Command
{ .\target\debug\pq2json.exe ..\..\Downloads\20200809201109710_53ea103e_e9fabfe6_001.parquet > 1 }

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 11
Milliseconds      : 361
Ticks             : 113610039
TotalDays         : 0.000131493100694444
TotalHours        : 0.00315583441666667
TotalMinutes      : 0.189350065
TotalSeconds      : 11.3610039
TotalMilliseconds : 11361.0039

PS C:\Users\mispecto\Projects\azure-kusto-parquet-conv> Measure-Command
{ .\target\debug\pq2json.exe --csv ..\..\Downloads\20200809201109710_53ea103e_e9fabfe6_001.parquet > 1 }

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 5
Milliseconds      : 624
Ticks             : 56247514
TotalDays         : 6.51012893518518E-05
TotalHours        : 0.00156243094444444
TotalMinutes      : 0.0937458566666667
TotalSeconds      : 5.6247514
TotalMilliseconds : 5624.7514
```
