+++ pq2json/Cargo.toml
@@ -12,3 +12,4 @@ serde = "1"
 serde_json = "1"
 num-bigint = "0.2"
 chrono = "0.4"
+csv = "1.1.4"
